refactor(arel): quote delegates to the connection like Rails
Arel::Visitors::ToSql#quote is a two-line delegation in Rails
(to_sql.rb:867-870) — SqlLiteral passes through, everything else goes to
@connection.quote. All value formatting (quoted_date, array encoding,
binary) lives in the adapter.

Trails' visitor owned a quotedDate/formatDate pair and the PostgreSQL
visitor overrode quote to encode array literals — neither exists in
Rails' Arel. Move that formatting onto the quoting hosts in
default-quoter.ts (which stand in for an adapter on the connection-less
Node#toSql debug path) and add postgresqlDefaultQuoter so the PG visitor
needs no quote override. With a real adapter, PG array literals now route
through the adapter's quote -> encode_array -> type_cast_array.
