+++ packages/arel/src/index.ts
@@ -9,7 +9,6 @@ export { DeleteManager } from "./delete-manager.js";
 import { TreeManager } from "./tree-manager.js";
 export { TreeManager };
 export { ArelError, EmptyJoinError, BindError } from "./errors.js";
-export { quoteArrayLiteral } from "./quote-array.js";
 export { relationName } from "./attributes/attribute.js";
 
 import { SqlLiteral } from "./nodes/sql-literal.js";
